data/pipelines/requirements.txt
requests==2.32.5
feedparser==6.0.12
supabase==2.7.4
google-generativeai==0.8.5
googlesearch-python==1.3.0
google-genai
pillow==12.0.0
python-dotenv
